labels hide when length > 3
