+++ src/client/components/containers/GraphContainer.tsx
@@ -1,5 +1,5 @@
 import * as React from "react";
-import BarGraph from "../display/BarGraph2.jsx";
+import BarGraph from "../display/BarGraph.jsx";
 
 export class GraphContainer extends React.Component<any, any> {
   constructor(props: any) {
